equivalence: report vacuous runs as inconclusive; fix two wrong kb decls
A differential where every seed agrees because neither side did any work
is not evidence, but it printed the same "N passed" line as a real run.
FUN_001c6900 reported "100 passed" at 5.7% coverage with every seed
bailing out of the same guard.

- unicorn_diff: after the pass path, classify a run as vacuous when
  coverage is below COVERAGE_FLOOR_PCT, or when >= 5 seeds all produced
  one observed behaviour (no return, scratch, or memory-write variation).
  New status "inconclusive", exit code 3, with a message pointing at
  --state-snapshot. A bare RET (<= 2 bytes) is exempt: no body to cover.
  The >= 5 seed floor keeps deliberately pinned single-input probes such
  as test_player_control_angle_endpoint out of the rule.
- lift_pipeline: handle status "inconclusive" like not_applicable
  (ran=False, gated on --equivalence-policy) so it can never satisfy a
  verification stage.
- batch_equivalence: label rc=3 INCONCLUSIVE rather than FAIL -- nothing
  diverged, but nothing was compared either.

kb decls corrected, both refuted by their own call sites:
- 0x138fd0 param 6 int -> float *out_rgb. Every caller passes a
  real_rgb_color: objects.c:4161 passes `float local_40[3]` and :3928 a
  float*, both through an (int) cast that existed only to satisfy the
  wrong decl; the sibling 0x1390d0 already declares float *out_rgb. The
  casts are removed at all three call sites.
- 0x231490 was named BinkSetSoundSystem(open_proc, close_proc), but its
  arguments at 0x1c61b4 are a __stdcall (U32)->void* allocator and a
  __stdcall (void*)->void deallocator -- RADSetMemory-shaped, not a
  sound-system registration. Renamed to the mechanical FUN_00231490
  with typed parameter names rather than guessing the RAD API name.

kb_meta rows reconciled against kb.json (3 of 8 contradicted it):
0x1aeb0 and 0x1b280 claimed status "ported" with no ported key in
kb.json and no implementation in src/ (only call sites); 0x1c0e0 claimed
"ported" while kb.json has ported:false and the address is in
deactivation_allowlist.json -- which hid a deliberate deactivation from
frontier.py and progress accounting while the deactivation gate stayed
quiet. All three set to unknown with the evidence recorded.

Also exported delinked/functions/000b7e30.obj (untracked dir) so
player_control_get_facing_angles resolves. Harness suite now 17 passed,
2 skipped, 0 failed -- the failure that forced --no-verify on 99c4ccb84
is gone. Build clean; vc71_regression over objects.c (215), objects/
object_lights.c (9) and bink/bink_playback.c (25) shows no regressions.

+++ src/halo/objects/objects.c
@@ -4114,7 +4114,7 @@ void FUN_0013a740(int param_1, int param_2, float *param_3)
       }
       iVar8 = CALL_FUN_001bf570(uVar7, 0, 0);
       if (iVar8 != 0) {
-        FUN_00138fd0(iVar5, uVar7, local_24, local_1c, local_14, (int)pfVar3);
+        FUN_00138fd0(iVar5, uVar7, local_24, local_1c, local_14, pfVar3);
       }
     }
   }
@@ -4344,7 +4344,7 @@ char FUN_0013ab20(unsigned int param_1, int param_2, int *param_3)
           (iVar4 = CALL_FUN_00138ee0(local_14), iVar4 != 0)) {
         CALL_FUN_001390d0(iVar6, local_14, puVar5, local_c, local_10,
                           (void *)local_88);
-        FUN_00138fd0(iVar6, iVar2, puVar5, local_c, local_10, (int)local_40);
+        FUN_00138fd0(iVar6, iVar2, puVar5, local_c, local_10, local_40);
         CALL_FUN_00180570((unsigned int)*puVar5 * 0x20 + *(int *)(iVar6 + 0xf8),
                           (void *)local_64);
         CALL_FUN_00180570((unsigned int)puVar5[1] * 0x20 +
